Feature #13271
Migrate Silverpeas Core to JakartaEE 10. For doing, a lot of changes have been done. The migration is used also to improve some codes.
* Because some field names in database tables are reserved words in H2, the execution of the table creation scripts fail with this RDBMS. So, those fields renamed and they are for all the supported RDBMS.
* Fix some JPQL in order to satisfy now the strict JPQL grammar expected by Hibernate (like in BasicJpaEntityRepository#deleteByComponentInstanceId(String))
* The beans that have to be managed by CDI require now to be annotated accordingly. The @Singleton annotation isn't taken as a managed bean annotation: those beans have to be annotated with a CDI stereotyped annotation.
* Remove the external social network support implementation like Facebook and Linkedin as they were provided by a deprecated and no-more-existing Spring project.
* Refactor I18n to be a class and a managed bean instead of an interface. I18NHelper doesn't extend it anymore but wraps it; its goal is to give a static access some of the I18n's methods.
* The war archive construction in the integration tests are adapted to the new Jakarta EE environment. It is rewritten in core-library.
* In core-api, reification of the component instance identity concept: the class modeling this concept takes in charge the knowledge how a component instance identifier is built according the nature of the component instance (personal, multi-user/multi-instanciable, ...)
* In some managed beans, replace the use of ServiceProvider to get a dependency by the use of an injection point (@Inject).
* Fix the SQL date/datetime <-> java date conversions in SqlTemporalParamSetter hierarchy classes.
* In order to ease the use of alternative User implementation, introduce in core-api the UserFactory and UserBuilder interfaces.
* The portlet implementation satisfying Jakarta EE is now carried by the Silverpeas Porlet project
* The CMIS support satisfying Jakarta EE is now carried by the Silverpeas CMIS projet
* Jackrabbit WebDAV support satisfying Jakarta EE is now carried by the Silverpeas Jackrabbit WebDAV project.
* Add a new interface FileItem and all classes extending the Apache FileItem implement now this interface. The Apache FileItem is now wrapped by the class DefaultFileItem.

+++ core-library/src/main/java/org/silverpeas/core/notification/user/delayed/DelayedNotificationManager.java
@@ -23,7 +23,6 @@
  */
 package org.silverpeas.core.notification.user.delayed;
 
-import com.ibm.icu.util.Calendar;
 import org.silverpeas.core.annotation.Service;
 import org.silverpeas.core.notification.user.client.NotificationManagerSettings;
 import org.silverpeas.core.notification.user.client.constant.NotifChannel;
@@ -38,15 +37,10 @@ import org.silverpeas.core.util.CollectionUtil;
 import org.silverpeas.core.util.DateUtil;
 import org.silverpeas.core.util.MapUtil;
 
-import javax.inject.Inject;
-import javax.transaction.Transactional;
-import java.util.Collection;
-import java.util.Date;
-import java.util.HashSet;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import jakarta.inject.Inject;
+import jakarta.transaction.Transactional;
+
+import java.util.*;
 
 /**
  * @author Yohann Chastagnier
@@ -150,7 +144,7 @@ public class DelayedNotificationManager implements DelayedNotification {
 
   @Override
   public int deleteDelayedNotifications(final Collection<Long> ids) {
-    int nbDeletes = 0;
+    long nbDeletes = 0;
     if (CollectionUtil.isNotEmpty(ids)) {
       for (final Collection<Long> idLot : CollectionUtil.split(ids)) {
         nbDeletes += dnRepository.deleteByIds(idLot);
@@ -158,7 +152,7 @@ public class DelayedNotificationManager implements DelayedNotification {
       nrRepository.deleteResources();
       dnRepository.flush();
     }
-    return nbDeletes;
+    return Math.toIntExact(nbDeletes);
   }
 
   /*
